feat(AzureAISDKv2): split client core class to small, service related parts.
