--- dotnet/src/Connectors/Connectors.AzureOpenAI/Core/AzureOpenAIClientCore.cs
@@ -1,101 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-
-using System;
-using System.Net.Http;
-using Azure.AI.OpenAI;
-using Azure.Core;
-using Microsoft.Extensions.Logging;
-using Microsoft.SemanticKernel.Services;
-
-namespace Microsoft.SemanticKernel.Connectors.AzureOpenAI;
-
-/// <summary>
-/// Core implementation for Azure OpenAI clients, providing common functionality and properties.
-/// </summary>
-internal sealed class AzureOpenAIClientCore : ClientCore
-{
-    /// <summary>
-    /// Gets the key used to store the deployment name in the <see cref="IAIService.Attributes"/> dictionary.
-    /// </summary>
-    public static string DeploymentNameKey => "DeploymentName";
-
-    /// <summary>
-    /// OpenAI / Azure OpenAI Client
-    /// </summary>
-    internal override AzureOpenAIClient Client { get; }
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="AzureOpenAIClientCore"/> class using API Key authentication.
-    /// </summary>
-    /// <param name="deploymentName">Azure OpenAI deployment name, see https://learn.microsoft.com/azure/cognitive-services/openai/how-to/create-resource</param>
-    /// <param name="endpoint">Azure OpenAI deployment URL, see https://learn.microsoft.com/azure/cognitive-services/openai/quickstart</param>
-    /// <param name="apiKey">Azure OpenAI API key, see https://learn.microsoft.com/azure/cognitive-services/openai/quickstart</param>
-    /// <param name="httpClient">Custom <see cref="HttpClient"/> for HTTP requests.</param>
-    /// <param name="logger">The <see cref="ILoggerFactory"/> to use for logging. If null, no logging will be performed.</param>
-    internal AzureOpenAIClientCore(
-        string deploymentName,
-        string endpoint,
-        string apiKey,
-        HttpClient? httpClient = null,
-        ILogger? logger = null) : base(logger)
-    {
-        Verify.NotNullOrWhiteSpace(deploymentName);
-        Verify.NotNullOrWhiteSpace(endpoint);
-        Verify.StartsWith(endpoint, "https://", "The Azure OpenAI endpoint must start with 'https://'");
-        Verify.NotNullOrWhiteSpace(apiKey);
-
-        var options = GetAzureOpenAIClientOptions(httpClient);
-
-        this.DeploymentOrModelName = deploymentName;
-        this.Endpoint = new Uri(endpoint);
-        this.Client = new AzureOpenAIClient(this.Endpoint, apiKey, options);
-    }
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="AzureOpenAIClientCore"/> class supporting AAD authentication.
-    /// </summary>
-    /// <param name="deploymentName">Azure OpenAI deployment name, see https://learn.microsoft.com/azure/cognitive-services/openai/how-to/create-resource</param>
-    /// <param name="endpoint">Azure OpenAI deployment URL, see https://learn.microsoft.com/azure/cognitive-services/openai/quickstart</param>
-    /// <param name="credential">Token credential, e.g. DefaultAzureCredential, ManagedIdentityCredential, EnvironmentCredential, etc.</param>
-    /// <param name="httpClient">Custom <see cref="HttpClient"/> for HTTP requests.</param>
-    /// <param name="logger">The <see cref="ILogger"/> to use for logging. If null, no logging will be performed.</param>
-    internal AzureOpenAIClientCore(
-        string deploymentName,
-        string endpoint,
-        TokenCredential credential,
-        HttpClient? httpClient = null,
-        ILogger? logger = null) : base(logger)
-    {
-        Verify.NotNullOrWhiteSpace(deploymentName);
-        Verify.NotNullOrWhiteSpace(endpoint);
-        Verify.StartsWith(endpoint, "https://", "The Azure OpenAI endpoint must start with 'https://'");
-
-        var options = GetAzureOpenAIClientOptions(httpClient);
-
-        this.DeploymentOrModelName = deploymentName;
-        this.Endpoint = new Uri(endpoint);
-        this.Client = new AzureOpenAIClient(this.Endpoint, credential, options);
-    }
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="AzureOpenAIClientCore"/> class using the specified OpenAIClient.
-    /// Note: instances created this way might not have the default diagnostics settings,
-    /// it's up to the caller to configure the client.
-    /// </summary>
-    /// <param name="deploymentName">Azure OpenAI deployment name, see https://learn.microsoft.com/azure/cognitive-services/openai/how-to/create-resource</param>
-    /// <param name="openAIClient">Custom <see cref="AzureOpenAIClient"/>.</param>
-    /// <param name="logger">The <see cref="ILogger"/> to use for logging. If null, no logging will be performed.</param>
-    internal AzureOpenAIClientCore(
-        string deploymentName,
-        AzureOpenAIClient openAIClient,
-        ILogger? logger = null) : base(logger)
-    {
-        Verify.NotNullOrWhiteSpace(deploymentName);
-        Verify.NotNull(openAIClient);
-
-        this.DeploymentOrModelName = deploymentName;
-        this.Client = openAIClient;
-
-        this.AddAttribute(DeploymentNameKey, deploymentName);
-    }
-}
